♻️ 8CH6NP task: split task intake resolve helpers
